Allow opening live mode also to iframe runtimes
